feature-extraction/pyproject.toml
[project]
name = "feature-extraction"
version = "0.1.0"
description = ""
authors = [
    {name = "goginesa",email = "sagar.gogineni@external.danone.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "urllib3 (==1.26.19)",
    "openai (>=1.59.7,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "langfuse (>=2.59.5,<3.0.0)"
]

[tool.poetry]
packages = [{include = "feature_extraction", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
